% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{day_dif}
\alias{day_dif}
\title{Get difference between daytime population and area population}
\usage{
day_dif(day, total)
}
\arguments{
\item{day}{A column representing the daytime population of a region}

\item{total}{A column representing the total population of a region}
}
\value{
A new column giving the difference between day and total
}
\description{
Get difference between daytime population and area population
}
