% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{facet_hist}
\alias{facet_hist}
\title{Generate facet histograms for each column in a dataframe}
\usage{
facet_hist(df, cols, title = "Histograms")
}
\arguments{
\item{df}{A data frame}

\item{cols}{A character vector of the columns you do not want faceted, such
as unique identifiers}

\item{title}{The title of the map plot (default is "Histograms").}
}
\value{
A ggplot object representing the histograms
}
\description{
This function takes a dataframe and the columns you want to keep, and creates
a faceted plot with histograms for all of the columns.
}
