% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips_data}
\alias{fips_data}
\title{Data from the Social Vulnerability Index}
\format{
\subsection{\code{fips_data}}{

A data frame with 72837 rows and 127 columns:
\describe{
\item{ST}{State-level FIPS code}
\item{STATE}{State}
\item{ST_ABBR}{State abbreviation}
\item{STCNTY}{County-level fips code}
\item{COUNTY}{FIPS}
...
}
}
}
\source{
\url{https://www.atsdr.cdc.gov/placeandhealth/svi/}
}
\usage{
fips_data
}
\description{
This data frame gives information relating to Social Vulnerability Index.
See \code{data_dict} df for more information on which each column represents
}
\keyword{datasets}
