% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{fix999}
\alias{fix999}
\title{Replace -999s with NAs for missing values}
\usage{
fix999(df)
}
\arguments{
\item{df}{A dataframe in which some cells have values of -999 to represent missing values}
}
\value{
The same dataframe but all -999s have been replaced with NA
}
\description{
Replace -999s with NAs for missing values
}
