% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{pct_day_dif}
\alias{pct_day_dif}
\title{Get percent change between daytime population and area population}
\usage{
pct_day_dif(day, total)
}
\arguments{
\item{day}{A column representing the daytime population of a region}

\item{total}{A column representing the total population of a region}
}
\value{
A new column giving the difference between day and total
}
\description{
This function accounts for when total pop is 0 - If there is any increase
in population during the day, we simplify to say it is a 100\% increase.
If day and total pop are both 0, we say percent change is 0.
}
