% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_hist}
\alias{plot_hist}
\title{Generate a histogram}
\usage{
plot_hist(df, col, title = "Histogram")
}
\arguments{
\item{df}{A data frame}

\item{col}{The column of the data frame to plot, as a string}

\item{title}{The title of the histogram plot (default is "Histogram").}
}
\value{
A ggplot object representing the histogram.
}
\description{
This function takes a numeric vector and creates a histogram using ggplot2.
}
