% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_map}
\alias{plot_map}
\title{Generate a colored map}
\usage{
plot_map(df, geometry, fill, title = "Map")

plot_map(df, geometry, fill, title = "Map")
}
\arguments{
\item{df}{A data frame}

\item{geometry}{The geometry to plot}

\item{fill}{The factor to color the map by - can be discrete or continuous}

\item{title}{The title of the map plot (default is "Map").}
}
\value{
A ggplot object representing the histogram.

A ggplot object representing the histogram.
}
\description{
This function takes a geometry object and fill scale creates a shaded map using ggplot2.

This function takes a geometry object and fill scale creates a shaded map using ggplot2.
}
