% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{resid_plot}
\alias{resid_plot}
\title{Residual Plot}
\usage{
resid_plot(lm1, title = "Residuals")
}
\arguments{
\item{lm1}{A linear model}

\item{title}{The title of the map plot (default is "Residuals").}
}
\value{
A ggplot object representing the residuals and fitted values
}
\description{
This function takes a linear model \code{lm} object and creates a residual plot
}
