% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{log_trans}
\alias{log_trans}
\title{Log transform columns}
\usage{
log_trans(df, cols)
}
\arguments{
\item{df}{A dataframe with columns to be transformed}

\item{cols}{The columns to be log_transformed, a \if{html}{\out{<tidy-select>}}}
}
\value{
A new dataframe with indicated columns log transformed
}
\description{
This function log transforms all of the indicated columns in a dataframe
and replaces the original columns with new ones titled \code{log_name}
}
